# proto/openapi.yaml
openapi: 3.1.0
info:
  title: SeaSnail OpenAPI
  version: 0.1.0
  description: |
    SeaSnail 本地音频转译服务的 HTTP/REST API。本地 UI 与第三方应用统一经此 API 接入（OpenAPI 为唯一集成面）。

    ## 鉴权
    Scoped Bearer Token，请求头 `Authorization: Bearer <token>`。token 独立于用户密码，可吊销、可多签、**绑 account_id**。
    可用 scope：
      - `sessions:read`    读取 / 搜索会话、取原始音频、导出
      - `sessions:write`  创建会话、编辑转译文本（**仅本地 UI token 可得，不对第三方签发**）
      - `sessions:delete` 删除会话（可按需签发给第三方）
      - `tokens:manage`    管理 token（**仅本地 UI token 可得**）

    **多账户模型**：MVP 支持多账户，每账户独立密码 + 独立 master DEK，账户间密钥完全隔离。日常免密只对**活跃账户**成立（其 DEK 缓存于 app 私有 keychain，`last_active_account_id` 记于非敏感 prefs 供重开取回）；切换账户需重新认证（`POST /accounts/{id}/unlock`）。

    - 首个账户：`POST /auth/setup`（免鉴权、仅 loopback、仅"未初始化"态可用）建首个账户并签发该账户 **root token**（`is_root=true` + `account_id`，持有全部 scope），root token secret 存 app 私有 keychain，本地 UI 据此调用。
    - 后续账户：`POST /accounts`（需任一已解锁账户的 root token）——无全局 root，靠"已解锁即可建新账户"（已确认）。
    - 切换/解锁：`POST /accounts/{id}/unlock`（免鉴权 + password）→ 解 wrap 该账户 DEK 入 keychain 作活跃 → 签发/恢复该账户 root token。
    - 第三方 token 经 `POST /tokens` 签发，**绑 account_id**（仅可访问该账户、且仅该账户已解锁时），**非 root 调用方不可被授予 `sessions:write` / `tokens:manage`**（daemon 硬约束），故第三方仅可得 `sessions:read` / `sessions:delete`。
    - 改密走 `POST /auth/password`（需 root token + 当前密码）。

    本地优先：服务监听本机；bootstrap 不走云端，token 由本地 UI 解锁后在 Integrations 面板生成。

    ## 数据安全
    入库数据（原始音频 + 转译内容）经加密落库（per-account 独立密钥）；读取 / 导出时由服务解密后返回。
...
    导出为敏感操作，需在请求体中提供用户密码再确认（作用于活跃账户）。
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ss_live_
  responses:
...
    Unauthorized:
      description: 未携带或无效 token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 权限不足（缺 scope / 密码错误）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: insufficient_scope }
            message: { type: string }
          required: [code, message]
      required: [error]
...
    Token:
      type: object
      properties:
        id: { type: string, format: uuid }
        account_id: { type: string, description: 所属账户（token 绑定此 account_id，仅可访问该账户且仅该账户已解锁时） }
        name: { type: string, description: 用户命名 }
        prefix: { type: string, example: ss_live_abcd…, description: 仅展示前缀 }
        is_root: { type: boolean, description: 是否 root token（每账户自己的 root，is_root + account_id；仅 setup/unlock 签发，持有 write/manage scope） }
        scopes:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time, nullable: true }
      required: [id, account_id, name, scopes, created_at]
    TokenCreate:
      type: object
      properties:
        name: { type: string }
        scopes:
          type: array
          items: { type: string }
          description: 从可用 scope 中选；非 root 调用方不可授予 sessions:write / tokens:manage（daemon 硬约束），仅 root 自身持有
      required: [name, scopes]
    TokenCreated:
      allOf:
        - $ref: '#/components/schemas/Token'
        - type: object
          properties:
            secret: { type: string, description: 完整 token 明文，仅创建时返回一次 }
          required: [secret]
...
security:
  - bearerAuth: []
paths:
...
  /tokens:
    get:
      summary: 列出 token
      x-required-scope: tokens:manage
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Token' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: 创建 token
      x-required-scope: tokens:manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenCreate' }
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenCreated' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tokens/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    delete:
      summary: 吊销 token
      x-required-scope: tokens:manage
      responses:
        '204': { description: 已吊销 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...